build: install Xvfb for headless test runs
