+++ package.json
@@ -183,7 +183,7 @@
     "vsce:publish:pre": "vsce publish --no-dependencies --pre-release",
     "compile-tests": "node -e \"try{require('fs').rmSync('out/test',{recursive:true,force:true})}catch(e){}\" && tsc -p tsconfig.test.json",
     "watch-tests": "tsc -p tsconfig.test.json -w",
-    "pretest": "npm run test:clean && npm run compile && npm run compile-tests && npm run lint",
+    "pretest": "npm run test:clean && npm run test:linux-deps && npm run compile && npm run compile-tests && npm run lint",
     "test": "node scripts/run-tests.js --scope=unit",
     "test:unit": "npm run pretest && node scripts/run-tests.js --scope=unit",
     "test:integration": "npm run pretest && node scripts/run-tests.js --scope=integration --install-deps --timeout=900000",
