+++ src/test/fetchApexLogHead.retry.test.ts
@@ -1,13 +1,8 @@
 import assert from 'assert/strict';
 import { EventEmitter } from 'events';
-import {
-  fetchApexLogHead,
-  OrgAuth,
-  __setExecFileImplForTests,
-  __resetExecFileImplForTests,
-  __setHttpsRequestImplForTests,
-  __resetHttpsRequestImplForTests
-} from '../salesforce';
+import { fetchApexLogHead, __setHttpsRequestImplForTests, __resetHttpsRequestImplForTests } from '../salesforce/http';
+import { __setExecFileImplForTests, __resetExecFileImplForTests } from '../salesforce/cli';
+import type { OrgAuth } from '../salesforce/types';
 
 suite('fetchApexLogHead retry', () => {
   teardown(() => {
